+++ spec/slack-gamebot/commands/leaderboard_spec.rb
@@ -11,6 +11,10 @@ describe SlackGamebot::Commands::Leaderboard do
       it 'displays leaderboard sorted by elo' do
         expect(message: "#{SlackRubyBot.config.user} leaderboard").to respond_with_slack_message "1. #{user_elo_48}\n2. #{user_elo_42}"
       end
+      it 'excludes unregistered users' do
+        user_elo_48.unregister!
+        expect(message: "#{SlackRubyBot.config.user} leaderboard").to respond_with_slack_message "1. #{user_elo_42}"
+      end
       it 'limits to max' do
         expect(message: "#{SlackRubyBot.config.user} leaderboard 1").to respond_with_slack_message "1. #{user_elo_48}"
       end